+++ src/features/usage-consolidation-accordion/summary-table/column-properties.js
@@ -1,5 +1,11 @@
+import React from 'react';
+import { FormattedNumber } from 'react-intl';
 
 import { combineMCLProps } from '../../../components/utilities';
+import {
+  formatValue,
+  formatCost,
+} from '../utilities';
 
 export const DEFAULT_SUMMARY_TABLE_COLUMNS = {
   COST: 'cost',
@@ -21,11 +27,20 @@ const DEFAULT_SUMMARY_TABLE_COLUMN_MAPPING = {
   [DEFAULT_SUMMARY_TABLE_COLUMNS.UC_ACTIONS]: null,
 };
 
-export const getSummaryTableColumnProperties = (intl, customProps = {}) => {
+const DEFAULT_FORMATTER = (currency) => ({
+  cost: rowData => formatValue(rowData.cost, (value) => formatCost(currency, value)),
+  costPerUse: rowData => formatValue(rowData.costPerUse, (value) => formatCost(currency, value)),
+  usage: rowData => formatValue(rowData.usage, (value) => <FormattedNumber value={value} />),
+});
+
+export const getCostPerUseFormatter = DEFAULT_FORMATTER;
+
+export const getSummaryTableColumnProperties = (intl, customProps = {}, currency) => {
   const defaultProps = {
     visibleColumns: [...Object.values(DEFAULT_SUMMARY_TABLE_COLUMNS)],
     columnMapping: { ...DEFAULT_SUMMARY_TABLE_COLUMN_MAPPING },
     columnWidths: { ...DEFAULT_SUMMARY_TABLE_COLUMN_WIDTH },
+    formatter: DEFAULT_FORMATTER(currency),
   };
 
   const combinedProps = combineMCLProps(defaultProps)(customProps);