+++ src/features/usage-consolidation-accordion/summary-table/summary-table.js
@@ -3,9 +3,7 @@ import PropTypes from 'prop-types';
 import {
   useIntl,
   FormattedMessage,
-  FormattedNumber,
 } from 'react-intl';
-import getSymbolFromCurrency from 'currency-symbol-map';
 
 import {
   MultiColumnList,
@@ -13,31 +11,44 @@ import {
   Dropdown,
   DropdownButton,
   DropdownMenu,
-  NoValue,
+  Icon,
   KeyValue,
 } from '@folio/stripes/components';
 
 import { getSummaryTableColumnProperties } from './column-properties';
-import { costPerUse as costPerUseShape } from '../../../constants';
+import {
+  costPerUse as costPerUseShape,
+  entityTypes,
+} from '../../../constants';
 
 const propTypes = {
   contentData: PropTypes.array,
   costPerUseData: costPerUseShape.CostPerUseReduxStateShape.isRequired,
+  costPerUseType: PropTypes.string.isRequired,
   customProperties: PropTypes.object,
+  entityType: PropTypes.string.isRequired,
   id: PropTypes.string.isRequired,
+  onViewTitles: PropTypes.func,
+  year: PropTypes.number.isRequired,
 };
 
 const SummaryTable = ({
   costPerUseData,
   customProperties,
   id,
+  year,
+  costPerUseType,
+  onViewTitles,
+  entityType,
   ...rest
 }) => {
   const intl = useIntl();
-  const { data } = costPerUseData;
+  const data = costPerUseData.data[costPerUseType];
+  if (!data) {
+    return null;
+  }
 
   const currency = data?.attributes?.parameters?.currency;
-  const currencySymbol = getSymbolFromCurrency(currency) || '';
 
   const {
     cost,
@@ -45,32 +56,15 @@ const SummaryTable = ({
     usage,
   } = data?.attributes?.analysis;
 
-  const formatCost = (value) => {
-    return (
-      <FormattedNumber value={value}>
-        {(formattedNumber) => `${currencySymbol}${formattedNumber} (${currency})`}
-      </FormattedNumber>
-    );
-  };
-
-  const formatValue = (value, callback) => {
-    const number = typeof value === 'string' ? Number(value) : value;
-
-    if (!number && number !== 0) {
-      return <NoValue />;
-    }
-
-    const valueToFixed = number.toFixed(2);
-
-    return callback ? callback(valueToFixed) : valueToFixed;
+  const handleViewTitles = (onToggle) => () => {
+    onViewTitles();
+    onToggle();
   };
 
   const formatter = {
-    cost: rowData => formatValue(rowData.cost, formatCost),
-    costPerUse: rowData => formatValue(rowData.costPerUse, formatCost),
-    usage: rowData => formatValue(rowData.usage, (value) => <FormattedNumber value={value} />),
     ucActions: () => (
       <Dropdown
+        id="summary-table-actions-dropdown"
         renderTrigger={({ onToggle, triggerRef, ariaProps, keyHandler, getTriggerProps }) => (
           <DropdownButton
             id="usage-consolidation-actions-dropdown-button"
@@ -88,18 +82,48 @@ const SummaryTable = ({
           <DropdownMenu
             role="menu"
           >
+            {entityType === entityTypes.PACKAGE ? (
+              <Button
+                id="summary-table-actions-view-titles"
+                buttonStyle="dropdownItem fullWidth"
+                role="menuitem"
+                onClick={handleViewTitles(onToggle)}
+                marginBottom0
+              >
+                <Icon
+                  icon="eye-open"
+                  size="small"
+                >
+                  <FormattedMessage id="ui-eholdings.usageConsolidation.summary.actions.view" />
+                </Icon>
+              </Button>
+            ) : null
+            }
             <Button
               buttonStyle="dropdownItem fullWidth"
               role="menuitem"
               onClick={onToggle}
               marginBottom0
             >
-              <FormattedMessage id="ui-eholdings.usageConsolidation.summary.actions.export" />
+              <Icon
+                icon="download"
+                size="small"
+              >
+                <FormattedMessage id="ui-eholdings.usageConsolidation.summary.actions.export" />
+              </Icon>
             </Button>
           </DropdownMenu>
         )}
       />
     )
+  };
+
+  const customPropertiesWithFormatter = {
+    ...customProperties,
+    formatter: {
+      ...(customProperties.formatter || {}),
+      ...formatter,
+    },
   };
 
   const contentData = rest.contentData || [{ cost, costPerUse, usage }];
@@ -109,15 +133,15 @@ const SummaryTable = ({
       <MultiColumnList
         id={id}
         contentData={contentData}
-        formatter={{
-          ...formatter,
-          ...customProperties.formatter,
-        }}
-        {...getSummaryTableColumnProperties(intl, customProperties)}
+        {...getSummaryTableColumnProperties(intl, customPropertiesWithFormatter, currency)}
         {...rest}
       />
     </KeyValue>
   );
+};
+
+SummaryTable.defaultProps = {
+  onViewTitles: () => {},
 };
 
 SummaryTable.propTypes = propTypes;