+++ src/redux/actions/costperuse/index.js
@@ -1,3 +1,6 @@
 export * from './get-cost-per-use';
 export * from './get-cost-per-use-success';
 export * from './get-cost-per-use-failure';
+export * from './get-cost-per-use-package-titles';
+export * from './get-cost-per-use-package-titles-success';
+export * from './get-cost-per-use-package-titles-failure';