+++ src/redux/reducers/costPerUse.js
@@ -2,6 +2,9 @@ import {
   GET_COST_PER_USE,
   GET_COST_PER_USE_SUCCESS,
   GET_COST_PER_USE_FAILURE,
+  GET_COST_PER_USE_PACKAGE_TITLES,
+  GET_COST_PER_USE_PACKAGE_TITLES_SUCCESS,
+  GET_COST_PER_USE_PACKAGE_TITLES_FAILURE,
 } from '../actions';
 
 import { formatErrors } from '../helpers';
@@ -19,7 +22,11 @@ const handleSuccess = (state, { payload }) => ({
   isLoading: false,
   isLoaded: true,
   isFailed: false,
-  data: payload,
+  data: {
+    ...state.data,
+    [payload.type]: payload,
+  },
+  errors: [],
 });
 
 const handlePendingRequest = (state) => ({
@@ -33,12 +40,51 @@ const handlers = {
   [GET_COST_PER_USE]: handlePendingRequest,
   [GET_COST_PER_USE_SUCCESS]: handleSuccess,
   [GET_COST_PER_USE_FAILURE]: handleError,
+  [GET_COST_PER_USE_PACKAGE_TITLES]: (state) => ({
+    ...state,
+    isPackageTitlesLoading: true,
+    isPackageTitlesLoaded: false,
+    isPackageTitlesFailed: false,
+  }),
+  [GET_COST_PER_USE_PACKAGE_TITLES_SUCCESS]: (state, { payload }) => ({
+    ...state,
+    isPackageTitlesLoading: false,
+    isPackageTitlesLoaded: true,
+    isPackageTitlesFailed: false,
+    data: {
+      ...state.data,
+      [payload.data.type]: payload.loadMore
+        ? {
+          ...payload.data,
+          attributes: {
+            ...payload.data.attributes,
+            resources: [
+              ...(state.data[payload.data.type]?.attributes?.resources || []),
+              ...payload.data.attributes.resources,
+            ],
+          },
+        }
+        : payload.data,
+    },
+    errors: [],
+  }),
+  [GET_COST_PER_USE_PACKAGE_TITLES_FAILURE]: (state, { payload }) => ({
+    ...state,
+    isPackageTitlesLoading: false,
+    isPackageTitlesLoaded: false,
+    isPackageTitlesFailed: true,
+    errors: formatErrors(payload.errors),
+  }),
 };
 
 const initialState = {
+  data: {},
   isLoading: false,
   isLoaded: false,
   isFailed: false,
+  isPackageTitlesLoading: false,
+  isPackageTitlesLoaded: false,
+  isPackageTitlesFailed: false,
   errors: [],
 };